Move design relevant components into LoadDesign component and  simplify landing page as router.
