+++ workbench/src/Components/LandingPage/LandingPage.js
@@ -1,15 +1,11 @@
-import React, {useCallback, useContext, useEffect, useRef, useState} from "react";
+import React, {useCallback} from "react";
 
-import {Trash} from "react-bootstrap-icons";
 import {LayoutManager} from "ui-layout-manager-dev";
 
-import splashScreen from "../../Assets/splash_screen.png";
 import layout from "../../layout.json";
-import {useWorkspace} from "../../Providers/GlobalProviders";
-import {useDalEngine} from "../../Providers/GlobalProviders";
-import ServerContext from "../../Providers/ServerContext";
 import {registry} from "../../Registry";
 import {useDesignLoaded} from "../../Store/useAppSelection";
+import {LoadDesign} from "./LoadDesign/LoadDesign";
 
 import "./LandingPage.scss";
 
@@ -18,58 +14,7 @@ import "./LandingPage.scss";
  * @return {JSX.Element}
  */
 export function LandingPage () {
-    const {workspace} = useWorkspace();
-    const [designs, setDesigns] = useState([]);
-    const [selectedDesign, setSelectedDesign] = useState(null);
-    const {sendJsonMessage} = useContext(ServerContext);
-    const [fileName, setFileName] = useState("");
-
     const designLoaded = useDesignLoaded();
-    const dalEngine = useDalEngine();
-
-    useEffect(() => {
-        if (workspace) {
-            setDesigns(workspace.map((item) => {
-                if (item.type === "file" && item.name.endsWith(".dal")) {
-                    return item;
-                }
-                return null;
-            }));
-        }
-    }, [workspace]);
-
-    const sendMessage = useCallback((type, payload) => {
-        sendJsonMessage({
-            type: type,
-            payload: payload,
-        });
-    }, [sendJsonMessage]);
-
-    const newDesign = useCallback((e) => {
-        e.stopPropagation();
-        e.preventDefault();
-        if (!fileName) return;
-        const fName = fileName.endsWith(".dal") ? fileName : fileName + ".dal";
-        sendMessage("create_design", {"fileName": fName});
-        setFileName("");
-    }, [fileName]);
-
-    const deleteDesign = useCallback((e) => {
-        e.stopPropagation();
-        e.preventDefault();
-        sendMessage("delete_design", {"fileName": selectedDesign.name});
-    }, [selectedDesign]);
-
-    const selectFile = (e, design) => {
-        e.stopPropagation();
-        setSelectedDesign(design);
-    };
-
-    const loadDesign = useCallback(() => {
-        if (!selectedDesign) return;
-        sendMessage("load_design", {"fileName": selectedDesign.name});
-    }, [selectedDesign]);
-
 
     const registryList = useCallback(() => registry, []);
 
@@ -78,52 +23,7 @@ export function LandingPage () {
             {
                 designLoaded?
                     <LayoutManager registry={registryList()} ldf={layout} /> :
-                    <div className="landing-page">
-                        <div className="splash">
-                            <div className="left">
-                                <img src={splashScreen} alt="Blueprint city" />
-                            </div>
-                            <div className="right">
-                                <div className="title-container">Design Workbench</div>
-                                <div className="create-design-row">
-                                    <input
-                                        className="file-name-input"
-                                        placeholder="Enter design name..."
-                                        value={fileName}
-                                        onChange={(e) => setFileName(e.target.value)}
-                                    />
-                                    <div className="create-btn" onClick={newDesign}>+ Create</div>
-                                </div>
-                                <div
-                                    className="file-selector-container"
-                                    onClick={(e) => selectFile(e, null)}>
-                                    <div className="files">
-                                        {designs.map((design) => (
-                                            <div key={design.uid}
-                                                onClick={(e) => selectFile(e, design)}
-                                                className="file"
-                                                style={design?.uid === selectedDesign?.uid ?
-                                                    {backgroundColor: "#3a4a5c"} : {}}>
-                                                {design.name}
-                                            </div>
-                                        ))}
-                                    </div>
-                                </div>
-                                <div className="button-row">
-                                    <div className="buttons-left">
-                                        <div className="icon-btn">
-                                            <Trash onClick={deleteDesign} />
-                                        </div>
-                                    </div>
-                                    <div className="buttons-right">
-                                        <span className="open-btn" onClick={loadDesign}>
-                                            Open Design
-                                        </span>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
+                    <LoadDesign />
             }
         </>
     );
