+++ workbench/src/Components/LandingPage/LoadDesign/LoadDesign.scss
@@ -1,0 +1,169 @@
+.landing-page {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    width: 100vw;
+    background: #1a1a1a;
+}
+
+.splash {
+    display: flex;
+    width: 800px;
+    height: 400px;
+    border-radius: 12px;
+    overflow: hidden;
+}
+
+.left {
+    flex: 1;
+}
+
+.left img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    filter: brightness(0.9) contrast(1.05);
+}
+
+.right {
+    flex: 1;
+    background: linear-gradient(135deg, #263244, #1b2430);
+    display: flex;
+    flex-direction: column;
+
+    .create-design-row {
+        display:flex;
+        flex-direction:row;
+        padding: 0 20px;
+        align-items: center;
+        margin-bottom: 10px;
+
+        .file-name-input {
+            flex:1;
+            font-size: 14px;
+            border: none;
+            background: none;
+            border: solid 1px #3d526f;
+            color:white;
+            padding: 3px;
+            border-radius: 0px;
+        }
+
+        .file-name-input::placeholder {
+            color: #576f92;
+            font-size:12px;
+        }
+
+        .create-btn {
+            font-family: "Inter", sans-serif;
+            color:#6a88b3;
+            display:flex;
+            align-items: center;
+            cursor: pointer;
+            padding:0 5px;
+
+            .btn-icon {
+                font-size: 24px;
+            }
+        }
+
+        .create-btn:hover {
+            color: #e6edf3;
+        }
+    }
+
+    .title-container {
+        font-family: "Inter", sans-serif;
+        font-size: 32px;
+        font-weight: 600;
+        color: #e6edf3;
+        letter-spacing: 0.5px;
+        height:80px;
+        margin-top: 10px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    }
+
+    .subtitle {
+        font-family: "Inter", sans-serif;
+        font-size: 14px;
+        font-weight: 400;
+        color: #9aa4b2;
+        text-align: left;
+        padding-left: 20px;
+        margin-bottom: 5px;
+    }
+
+    .file-selector-container {
+        flex: 1;
+        margin: 0 20px;
+        display:flex;
+        flex-direction: column;
+        overflow-y: auto;
+        scrollbar-gutter: stable;
+        scrollbar-color: #47474766 #252526;
+        scrollbar-width: thin;
+        font-size:14px;
+        border: solid 1px #2a3948;
+
+        .files {
+            flex: 1;
+            flex-direction: column;
+
+            .file {
+                font-family: "Inter", sans-serif;
+                font-size: 13px;
+                font-weight: 500;
+                color: #e6edf3;
+                padding: 10px 20px 10px 10px;
+                cursor: pointer;
+                transition: background-color 0.2s ease;
+
+                &:hover {
+                    background-color: #3a4a5c;
+                }
+            }
+        }
+    }
+
+    .button-row {
+        margin-top:5px;
+        padding-right:15px;
+        margin-bottom: 10px;
+        display: flex;
+        flex-direction: row;
+        justify-content:space-between;
+        height:25px;
+
+        .buttons-left {
+            display: flex;
+            flex-direction: row;
+            padding-left:20px;
+            .icon-btn{
+                padding:5px;
+                font-family: "Inter", sans-serif;
+                color:rgb(196, 196, 196);
+                cursor: pointer;
+            }
+            .icon-btn:hover{
+                color:white;
+            }
+        }
+        
+        .buttons-right {
+            display: flex;
+            flex-direction: row;
+            .open-btn{
+                padding:5px;
+                font-family: "Inter", sans-serif;
+                color:rgb(196, 196, 196);
+                cursor: pointer;
+            }
+            .open-btn:hover{
+                color:white;
+            }   
+        }
+    }
+}
